fix(R12): use crypto.randomUUID() for markdown code block placeholders
convertMarkdownBold() used predictable sequential placeholders
(\x00CODE0\x00, \x00CODE1\x00, ...) when protecting code blocks from
bold conversion. An attacker who includes these exact strings in a
message can hijack the placeholder restoration step:
- \x00CODE0\x00 in attacker text gets replaced with the first code
  block's content, displacing it from its original position
- Remaining code blocks' placeholders are left as raw null bytes

Extracts markdown utilities to a separate module (markdown.ts) for
testability and replaces the predictable counter with
crypto.randomUUID(), making placeholder collision infeasible.

Addresses: F-12 (predictable placeholder collision in MarkdownV2)
